Updating tests folder, little layout adjustments, uploading screenshots and updating readme.
