+++ package.json
@@ -51,11 +51,11 @@
     ]
   },
   "devDependencies": {
-    "chai": "4.2.0",
     "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "1.15.2",
     "enzyme-to-json": "3.4.4",
-    "supertest": "4.0.2"
+    "supertest": "4.0.2",
+    "chai": "4.2.0"
   },
   "jest": {
     "snapshotSerializers": [
