+++ src/App.scss
@@ -13,6 +13,7 @@ body {
 .logo {
   margin: 24px auto 10px;
   width: 140px;
+  border: 1px solid #ffffff;
 }
 
 .infinite-container {
@@ -24,7 +25,7 @@ body {
   bottom: 40px;
   width: auto;
   left: 47%;
-  background: #ffffff;
+  background: red;
   border-radius: 50%;
   padding: 20px;
   -webkit-box-shadow: 0px 0px 20px -1px rgba(0, 0, 0, 0.75);
